# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Food & Handmade Products Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RegisterBodyRole = typeof RegisterBodyRole[keyof typeof RegisterBodyRole];


export const RegisterBodyRole = {
  buyer: 'buyer',
  seller: 'seller',
} as const;

export interface RegisterBody {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  phone?: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  buyer: 'buyer',
  seller: 'seller',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  sellerApproved: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Address {
  id: number;
  userId: number;
  name: string;
  street: string;
  city: string;
  state: string;
  zipCode: string;
  phone: string;
  isDefault: boolean;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: string;
  sellerApproved: boolean;
  addresses: Address[];
  createdAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  phone?: string;
}

export interface AddressBody {
  name: string;
  street: string;
  city: string;
  state: string;
  zipCode: string;
  phone: string;
  isDefault?: boolean;
}

export interface Product {
  id: number;
  sellerId: number;
  name: string;
  category: string;
  price: number;
  stock: number;
  description: string;
  images: string[];
  unit: string;
  tags: string[];
  sellerName: string;
  shopName: string;
  createdAt: string;
}

export interface ProductDetail {
  id: number;
  sellerId: number;
  name: string;
  category: string;
  price: number;
  stock: number;
  description: string;
  images: string[];
  unit: string;
  tags: string[];
  sellerName: string;
  shopName: string;
  sellerPhone: string;
  sellerAddress: string;
  sellerWhatsapp?: string;
  orderingMode: string;
  createdAt: string;
}

export interface ProductBody {
  name: string;
  category: string;
  price: number;
  stock: number;
  description: string;
  images?: string[];
  unit: string;
  tags?: string[];
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  subtotal: number;
  unit: string;
}

export interface DeliveryAddress {
  name: string;
  street: string;
  city: string;
  state: string;
  zipCode: string;
  phone: string;
}

export type OrderOrderMode = typeof OrderOrderMode[keyof typeof OrderOrderMode];


export const OrderOrderMode = {
  ecommerce: 'ecommerce',
  whatsapp: 'whatsapp',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Order {
  id: number;
  orderId: string;
  buyerId: number;
  sellerId: number;
  buyerName: string;
  sellerName: string;
  shopName: string;
  items: OrderItem[];
  totalPrice: number;
  deliveryAddress: DeliveryAddress;
  orderMode: OrderOrderMode;
  status: OrderStatus;
  paymentMethod?: string;
  paymentStatus: OrderPaymentStatus;
  sellerNotes?: string;
  trackingNumber?: string;
  createdAt: string;
  updatedAt: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
};

export type CreateOrderBodyOrderMode = typeof CreateOrderBodyOrderMode[keyof typeof CreateOrderBodyOrderMode];


export const CreateOrderBodyOrderMode = {
  ecommerce: 'ecommerce',
  whatsapp: 'whatsapp',
} as const;

export interface CreateOrderBody {
  sellerId: number;
  items: CreateOrderBodyItemsItem[];
  deliveryAddress: DeliveryAddress;
  orderMode: CreateOrderBodyOrderMode;
  paymentMethod?: string;
}

export type UpdateOrderStatusBodyStatus = typeof UpdateOrderStatusBodyStatus[keyof typeof UpdateOrderStatusBodyStatus];


export const UpdateOrderStatusBodyStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  sellerNotes?: string;
  trackingNumber?: string;
}

export interface CancelOrderBody {
  reason?: string;
}

export type SellerProfileOrderingMode = typeof SellerProfileOrderingMode[keyof typeof SellerProfileOrderingMode];


export const SellerProfileOrderingMode = {
  ecommerce: 'ecommerce',
  whatsapp: 'whatsapp',
} as const;

export interface SellerProfile {
  id: number;
  userId: number;
  shopName: string;
  phone: string;
  address: string;
  orderingMode: SellerProfileOrderingMode;
  whatsappNumber?: string;
  paymentMethods: string[];
  verified: boolean;
  userName: string;
  userEmail: string;
  totalProducts?: number;
  totalOrders?: number;
  createdAt: string;
}

export type SellerProfileBodyOrderingMode = typeof SellerProfileBodyOrderingMode[keyof typeof SellerProfileBodyOrderingMode];


export const SellerProfileBodyOrderingMode = {
  ecommerce: 'ecommerce',
  whatsapp: 'whatsapp',
} as const;

export interface SellerProfileBody {
  shopName: string;
  phone: string;
  address: string;
  orderingMode: SellerProfileBodyOrderingMode;
  whatsappNumber?: string;
  paymentMethods?: string[];
}

export interface SellerDashboard {
  totalProducts: number;
  totalOrders: number;
  pendingOrders: number;
  totalRevenue: number;
  recentOrders: Order[];
  orderingMode: string;
}

export interface SalesHistory {
  totalSales: number;
  totalOrders: number;
  avgOrderValue: number;
  orders: Order[];
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
  stock: number;
  sellerId: number;
  sellerName: string;
  unit: string;
  subtotal: number;
}

export interface Cart {
  id: number;
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  productId: number;
  /** @minimum 1 */
  quantity: number;
}

export interface UpdateCartItemBody {
  /** @minimum 0 */
  quantity: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
sort?: ListProductsSort;
page?: number;
limit?: number;
sellerId?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  popular: 'popular',
} as const;

export type ListOrdersParams = {
status?: string;
page?: number;
};

export type GetSellerOrdersParams = {
status?: string;
page?: number;
};

export type GetSellerSalesHistoryParams = {
dateFrom?: string;
dateTo?: string;
};