# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food & Handmade Products Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sellerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (seller only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product (seller only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product (seller only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted

  # ORDERS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    # Auth
...
    Product:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        unit:
          type: string
        tags:
          type: array
          items:
            type: string
        sellerName:
          type: string
        shopName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sellerId
        - name
        - category
        - price
        - stock
        - description
        - images
        - unit
        - tags
        - sellerName
        - shopName
        - createdAt
    ProductDetail:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        unit:
          type: string
        tags:
          type: array
          items:
            type: string
        sellerName:
          type: string
        shopName:
          type: string
        sellerPhone:
          type: string
        sellerAddress:
          type: string
        sellerWhatsapp:
          type: string
        orderingMode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - sellerId
        - name
        - category
        - price
        - stock
        - description
        - images
        - unit
        - tags
        - sellerName
        - shopName
        - sellerPhone
        - sellerAddress
        - orderingMode
        - createdAt
    ProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        price:
          type: number
        stock:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        unit:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - category
        - price
        - stock
        - description
        - unit
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - limit
        - totalPages

    # Orders
...